backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
google-generativeai==0.3.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
rdkit==2022.9.5
deepchem==2.8.0
reportlab==4.0.9
chromadb==0.4.21
requests==2.31.0
biopython==1.81
numpy==1.24.3
scipy==1.11.4
pandas==2.1.1
scikit-learn==1.3.2
torch>=2.0.0
tensorflow>=2.11.0
